Refactor: Extracts data picker view model from daily report to make that class easier to understand
