+++ app/models/view/show_daily_report.rb
@@ -41,7 +41,7 @@ module View
 
     def available_dates
       dates_raw.map do |date|
-        DatePresenter.new(date, querier.date)
+        ReportDatePicker.new(date, querier.date)
       end
     end
 
@@ -64,22 +64,6 @@ module View
 
     def calculations
       @calclations ||= DailyCalculations.new(data)
-    end
-
-    class DatePresenter
-      attr_reader :formatter, :is_selected
-
-      def initialize(date, selected_date)
-        @is_selected = date == selected_date
-        @formatter = DateTimeFormatter.new(date)
-      end
-
-      delegate :database_date, :full_month_us_date,
-        to: :formatter
-
-      def selected
-        is_selected ? 'selected' : ''
-      end
     end
 
     def format_date(raw_date)
